==> analysis/4_DE_gene_comparison_with_6month_brain_and_validating_qPCR_plot.Rmd <==
---
title: "4_DE gene comparison with 6month brain and validating qPCR plot"
author: "Yang & Steve"
date: "24/02/2020"
output: 
  html_document: 
    toc: yes
    toc_float: yes
    code_folding: hide
editor_options: 
  chunk_output_type: console
---
### Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

```{r packages}
library(limma)
library(edgeR)
library(tidyverse)
library(magrittr)
library(pander)
library(ggrepel)
library(scales)
library(plyr)
library(ggraph)
library(tidygraph)
library(readxl)
theme_set(theme_bw())
panderOptions("big.mark", ",")
panderOptions("table.split.table", Inf)
panderOptions("table.style", "rmarkdown")
if (interactive()) setwd(here::here("analysis"))
```

# Data load

```{r loaddgeList}
dgeList <- read_rds(here::here("data","dgeList.rds"))
entrezGenes <- dgeList$genes %>%
  dplyr::filter(!is.na(entrez_gene)) %>%
  unnest(entrez_gene) %>%
  dplyr::rename(entrez_gene = entrez_gene)
topTable <- file.path(here::here("output", "topTable.csv")) %>% 
  read_csv()
sixmonthtopTable <- read_excel(here::here("files", "6monthDEgenes.xlsx"))
```

# Find common DE genes
```{r common DE genes}
commonDE <- topTable %>%
  dplyr::filter(FDR < 0.01) %>%
  dplyr::filter(ensembl_gene_id %in% sixmonthtopTable$`Ensembl ID`) %>%
  dplyr::select(ensembl_gene_id, 
                external_gene_name)
commonDE %>%
  pander(
      style = "rmarkdown", 
      split.tables = Inf,
      caption = "Common DE genes in 7 dpf larvae and 6 month brain.",
      justify = "lr"
  )
```

# Plot expression of DE genes for validating qPCR

Rpl13 gene was used as a house-keeping genes. The expression level is similar across all samples.

```{r plotrpl13, echo=FALSE, fig.cap = "Expression of *rpl13* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000099380",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plotCABZ01034698, echo=FALSE, fig.cap = "Expression of *CABZ01034698* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000099511",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plotmdh1ab, echo=FALSE, fig.cap = "Expression of *mdh1ab* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000103849",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plotsi:ch211-11p18.6, echo=FALSE, fig.cap = "Expression of *si:ch211-11p18.6* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000077068",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plotsi:ch211-213a13.2, echo=FALSE, fig.cap = "Expression of *si:ch211-213a13.2* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000093024",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plottfa, echo=FALSE, fig.cap = "Expression of *tfa* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000016771",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plottfr1b, echo=FALSE, fig.cap = "Expression of *tfr1b* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000077372",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```

```{r plotfthl31, echo=FALSE, fig.cap = "Expression of *fthl31* across all samples."}
dgeList %>%
	cpm(log = TRUE) %>%
	.["ENSDARG00000094210",] %>%
	as.matrix() %>%
	set_colnames("logCPM") %>%
	cbind(dgeList$samples[rownames(.),]) %>%
	ggplot(aes(pair, logCPM, fill = Genotype)) +
	geom_bar(stat = "identity") +
  facet_wrap(~Genotype, scales = "free_x") +
  guides(fill = FALSE) 
```
